Implement swap feature.
Swap button now creates a list of transactions for the user to sign and send (only non-intents transactions so far).
Because of slight formatting differences, added new interfaces and helper functions to construct the transactions returned by the DEX aggregator.
Also handles unwrapping NEAR if `needs_unwrap` is true, using the difference in user's wNEAR balance before and after the swap.